backend/evals/Semantic_Retreival_Auto_Eval/pyproject.toml
[project]
name = "poetry-tester"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikhileshwarreddybommareddy",email = "bnikhileshwarreddy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "llama-index-llms-openai (>=0.3.28,<0.4.0)",
    "llama-index-core (>=0.12.26,<0.13.0)",
    "llama-index-embeddings-huggingface (>=0.5.2,<0.6.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
